remove cloud provider dependencies to pkg/volume
