+++ pkg/cloudprovider/providers/gce/gce_disks.go
@@ -29,7 +29,9 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/sets"
 	cloudprovider "k8s.io/cloud-provider"
-	"k8s.io/kubernetes/pkg/volume"
+	cloudvolume "k8s.io/cloud-provider/volume"
+	volerr "k8s.io/cloud-provider/volume/errors"
+	volumehelpers "k8s.io/cloud-provider/volume/helpers"
 	volumeutil "k8s.io/kubernetes/pkg/volume/util"
 
 	"github.com/GoogleCloudPlatform/k8s-cloud-provider/pkg/cloud"
@@ -504,7 +506,7 @@ func newDiskMetricContextRegional(request, region string) *metricContext {
 // GetLabelsForVolume retrieved the label info for the provided volume
 func (g *Cloud) GetLabelsForVolume(ctx context.Context, pv *v1.PersistentVolume) (map[string]string, error) {
 	// Ignore any volumes that are being provisioned
-	if pv.Spec.GCEPersistentDisk.PDName == volume.ProvisionedVolumeName {
+	if pv.Spec.GCEPersistentDisk.PDName == cloudvolume.ProvisionedVolumeName {
 		return nil, nil
 	}
 
@@ -730,7 +732,7 @@ func getDiskType(diskType string) (string, error) {
 func (g *Cloud) DeleteDisk(diskToDelete string) error {
 	err := g.doDeleteDisk(diskToDelete)
 	if isGCEError(err, "resourceInUseByAnotherResource") {
-		return volume.NewDeletedVolumeInUseError(err.Error())
+		return volerr.NewDeletedVolumeInUseError(err.Error())
 	}
 
 	if err == cloudprovider.DiskNotFound {
@@ -811,7 +813,7 @@ func (g *Cloud) GetAutoLabelsForPD(name string, zone string) (map[string]string,
 		// However it is more consistent to ensure the disk exists,
 		// and in future we may gather addition information (e.g. disk type, IOPS etc)
 		if utilfeature.DefaultFeatureGate.Enabled(cloudfeatures.GCERegionalPersistentDisk) {
-			zoneSet, err := volumeutil.LabelZonesToSet(zone)
+			zoneSet, err := volumehelpers.LabelZonesToSet(zone)
 			if err != nil {
 				klog.Warningf("Failed to parse zone field: %q. Will use raw field.", zone)
 			}
@@ -852,7 +854,7 @@ func (g *Cloud) GetAutoLabelsForPD(name string, zone string) (map[string]string,
 			return nil, fmt.Errorf("PD is regional but does not have any replicaZones specified: %v", disk)
 		}
 		labels[v1.LabelZoneFailureDomain] =
-			volumeutil.ZonesSetToLabelValue(zoneInfo.replicaZones)
+			volumehelpers.ZonesSetToLabelValue(zoneInfo.replicaZones)
 		labels[v1.LabelZoneRegion] = disk.Region
 	case nil:
 		// Unexpected, but sanity-check
