+++ staging/publishing/import-restrictions.yaml
@@ -190,6 +190,7 @@
   - k8s.io/apimachinery
   - k8s.io/apiserver
   - k8s.io/client-go
+  - k8s.io/cloud-provider
   - k8s.io/klog
   - k8s.io/utils
 
@@ -213,4 +214,5 @@
   - k8s.io/api
   - k8s.io/apimachinery
   - k8s.io/klog
+  - k8s.io/cloud-provider
   - k8s.io/csi-translation-lib
