+++ staging/publishing/rules.yaml
@@ -989,3 +989,5 @@ rules:
       branch: master
     - repository: apimachinery
       branch: master
+    - repository: cloud-provider
+      branch: master
