+++ staging/src/k8s.io/cloud-provider/BUILD
@@ -38,6 +38,7 @@ filegroup(
         "//staging/src/k8s.io/cloud-provider/features:all-srcs",
         "//staging/src/k8s.io/cloud-provider/node:all-srcs",
         "//staging/src/k8s.io/cloud-provider/service/helpers:all-srcs",
+        "//staging/src/k8s.io/cloud-provider/volume:all-srcs",
     ],
     tags = ["automanaged"],
 )
