+++ staging/src/k8s.io/cloud-provider/volume/helpers/BUILD
@@ -1,20 +1,26 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "aws_ebs.go",
-        "gce_pd.go",
-        "in_tree_volume.go",
-        "openstack_cinder.go",
-    ],
-    importmap = "k8s.io/kubernetes/vendor/k8s.io/csi-translation-lib/plugins",
-    importpath = "k8s.io/csi-translation-lib/plugins",
+    srcs = ["zones.go"],
+    importmap = "k8s.io/kubernetes/vendor/k8s.io/cloud-provider/volume/helpers",
+    importpath = "k8s.io/cloud-provider/volume/helpers",
     visibility = ["//visibility:public"],
     deps = [
         "//staging/src/k8s.io/api/core/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/util/sets:go_default_library",
         "//staging/src/k8s.io/cloud-provider/volume:go_default_library",
+        "//vendor/k8s.io/klog:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = ["zones_test.go"],
+    embed = [":go_default_library"],
+    deps = [
+        "//staging/src/k8s.io/api/core/v1:go_default_library",
+        "//staging/src/k8s.io/apimachinery/pkg/util/sets:go_default_library",
     ],
 )
